requirements.txt
streamlit>=1.28.0
reportlab>=4.0.0
streamlit-authenticator>=0.4.1
PyYAML>=6.0
